Move wasmer helper functions to helpers.go
